feat: update categorie and product get method

+++ BcCardapioDigital.API/Domain/Repositories/IRepositorioCategoria.cs
@@ -5,7 +5,7 @@ namespace BcCardapioDigital.API.Domain.Repositories
     public interface IRepositorioCategoria
     {
         Task<bool> CriarCategoria(Categoria entity);
-        Task<Categoria?> BuscarCategoria(long id);
+        Task<Categoria?> BuscarCategoria(int id);
         Task<List<Categoria>> ListarCategorias();
         Task<bool> RemoverCategoria(Categoria entity);
         Task<bool> Atualizar(Categoria entity);
