+++ BcCardapioDigital.API/Domain/Repositories/IRepositorioProduto.cs
@@ -5,7 +5,7 @@ namespace BcCardapioDigital.API.Domain.Repositories
     public interface IRepositorioProduto
     {
         Task<bool> CriarProduto(Produto entity);
-        Task<Produto?> BuscarProduto(long id);
+        Task<Produto?> BuscarProduto(int id);
         Task<List<Produto>> ListarProdutos();
         Task<List<Produto>> ProdutosPopulares();
         Task<bool> RemoverProduto(Produto entity);
